More description on single estate view
